Extract and fix query tests
