+++ src/lib.rs
@@ -21,11 +21,17 @@ mod unittest_cw721_revoke_all;
 mod unittest_helpers;
 mod unittest_make_ownership_private;
 mod unittest_mint;
+mod unittest_minters;
 mod unittest_private_metadata;
 mod unittest_public_metadata;
 mod unittest_queries;
 mod unittest_reveal;
 mod unittest_transfer;
+mod unittest_query_contract_config;
+mod unittest_query_contract_info;
+mod unittest_query_minters;
+mod unittest_query_all_tokens;
+mod unittest_query_num_tokens;
 mod unittest_register_receive_nft;
 mod unittest_send;
 mod unittest_set_global_approval;
