+++ src/unittest_queries.rs
@@ -13,488 +13,7 @@ mod tests {
     use cosmwasm_std::{
         from_binary, BlockInfo, Env, HumanAddr, MessageInfo,
     };
-    use crate::unittest_helpers::{extract_error_msg, init_helper_default, init_helper_with_config};
-
-    // test ContractInfo query
-    #[test]
-    fn test_query_contract_info() {
-        let (init_result, deps) = init_helper_default();
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-        let query_msg = QueryMsg::ContractInfo {};
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::ContractInfo { name, symbol } => {
-                assert_eq!(name, "sec721".to_string());
-                assert_eq!(symbol, "S721".to_string());
-            }
-            _ => panic!("unexpected"),
-        }
-    }
-
-    // test ContractConfig query
-    #[test]
-    fn test_query_contract_config() {
-        let (init_result, deps) =
-            init_helper_with_config(false, true, true, false, true, false, true);
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-        let query_msg = QueryMsg::ContractConfig {};
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::ContractConfig {
-                token_supply_is_public,
-                owner_is_public,
-                sealed_metadata_is_enabled,
-                unwrapped_metadata_is_private,
-                minter_may_update_metadata,
-                owner_may_update_metadata,
-                burn_is_enabled,
-            } => {
-                assert_eq!(token_supply_is_public, false);
-                assert_eq!(owner_is_public, true);
-                assert_eq!(sealed_metadata_is_enabled, true);
-                assert_eq!(unwrapped_metadata_is_private, false);
-                assert_eq!(minter_may_update_metadata, true);
-                assert_eq!(owner_may_update_metadata, false);
-                assert_eq!(burn_is_enabled, true);
-            }
-            _ => panic!("unexpected"),
-        }
-    }
-
-    // test minters query
-    #[test]
-    fn test_query_minters() {
-        let (init_result, mut deps) = init_helper_default();
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-        let alice = HumanAddr("alice".to_string());
-        let bob = HumanAddr("bob".to_string());
-        let charlie = HumanAddr("charlie".to_string());
-
-        let minters = vec![
-            alice.clone(),
-            bob.clone(),
-            charlie.clone(),
-            bob.clone(),
-            alice.clone(),
-        ];
-        let handle_msg = HandleMsg::SetMinters {
-            minters: minters.clone(),
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-
-        let query_msg = QueryMsg::Minters {};
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::Minters { minters } => {
-                assert_eq!(minters.len(), 3);
-                assert!(minters.contains(&alice));
-                assert!(minters.contains(&bob));
-                assert!(minters.contains(&charlie));
-            }
-            _ => panic!("unexpected"),
-        }
-    }
-
-    // test NumTokens query
-    #[test]
-    fn test_query_num_tokens() {
-        let (init_result, mut deps) =
-            init_helper_with_config(false, true, true, false, true, false, true);
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT1".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My1".to_string()),
-                description: Some("Public 1".to_string()),
-                image: Some("URI 1".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT2".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My2".to_string()),
-                description: Some("Public 2".to_string()),
-                image: Some("URI 2".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-
-        // test non-minter attempt
-        let alice = HumanAddr("alice".to_string());
-        let bob = HumanAddr("bob".to_string());
-        let charlie = HumanAddr("charlie".to_string());
-
-        let minters = vec![
-            alice.clone(),
-            bob.clone(),
-            charlie.clone(),
-            bob.clone(),
-            alice.clone(),
-        ];
-        let handle_msg = HandleMsg::SetMinters {
-            minters: minters.clone(),
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-
-        let query_msg = QueryMsg::NumTokens { viewer: None };
-        let query_result = query(&deps, query_msg);
-        let error = extract_error_msg(query_result);
-        assert!(error.contains("The token supply of this contract is private"));
-
-        // test minter with bad viewing key
-        let viewer = ViewerInfo {
-            address: alice.clone(),
-            viewing_key: "key".to_string(),
-        };
-        let query_msg = QueryMsg::NumTokens {
-            viewer: Some(viewer.clone()),
-        };
-        let query_result = query(&deps, query_msg);
-        let error = extract_error_msg(query_result);
-        assert!(error.contains("Wrong viewing key for this address or viewing key not set"));
-
-        // test valid minter, valid key
-        let handle_msg = HandleMsg::SetViewingKey {
-            key: "key".to_string(),
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("alice", &[]), handle_msg);
-        let query_msg = QueryMsg::NumTokens {
-            viewer: Some(viewer.clone()),
-        };
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::NumTokens { count } => {
-                assert_eq!(count, 2);
-            }
-            _ => panic!("unexpected"),
-        }
-
-        // test token supply public
-        let (init_result, mut deps) =
-            init_helper_with_config(true, true, true, false, true, false, true);
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT1".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My1".to_string()),
-                description: Some("Public 1".to_string()),
-                image: Some("URI 1".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT2".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My2".to_string()),
-                description: Some("Public 2".to_string()),
-                image: Some("URI 2".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let viewer = ViewerInfo {
-            address: alice.clone(),
-            viewing_key: "key".to_string(),
-        };
-        let query_msg = QueryMsg::NumTokens {
-            viewer: Some(viewer.clone()),
-        };
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::NumTokens { count } => {
-                assert_eq!(count, 2);
-            }
-            _ => panic!("unexpected"),
-        }
-    }
-
-    // test AllTokens query
-    #[test]
-    fn test_query_all_tokens() {
-        let (init_result, mut deps) =
-            init_helper_with_config(false, true, true, false, true, false, true);
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT1".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My1".to_string()),
-                description: Some("Public 1".to_string()),
-                image: Some("URI 1".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT2".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My2".to_string()),
-                description: Some("Public 2".to_string()),
-                image: Some("URI 2".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT3".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My3".to_string()),
-                description: Some("Public 3".to_string()),
-                image: Some("URI 3".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-
-        // test non-minter attempt
-        let alice = HumanAddr("alice".to_string());
-        let bob = HumanAddr("bob".to_string());
-        let charlie = HumanAddr("charlie".to_string());
-
-        let minters = vec![
-            alice.clone(),
-            bob.clone(),
-            charlie.clone(),
-            bob.clone(),
-            alice.clone(),
-        ];
-        let handle_msg = HandleMsg::SetMinters {
-            minters: minters.clone(),
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-
-        let query_msg = QueryMsg::AllTokens {
-            viewer: None,
-            start_after: None,
-            limit: None,
-        };
-        let query_result = query(&deps, query_msg);
-        let error = extract_error_msg(query_result);
-        assert!(error.contains("The token supply of this contract is private"));
-
-        // test minter with bad viewing key
-        let viewer = ViewerInfo {
-            address: alice.clone(),
-            viewing_key: "key".to_string(),
-        };
-        let query_msg = QueryMsg::AllTokens {
-            viewer: Some(viewer.clone()),
-            start_after: None,
-            limit: None,
-        };
-        let query_result = query(&deps, query_msg);
-        let error = extract_error_msg(query_result);
-        assert!(error.contains("Wrong viewing key for this address or viewing key not set"));
-
-        // test valid minter, valid key only return first two
-        let handle_msg = HandleMsg::SetViewingKey {
-            key: "key".to_string(),
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("alice", &[]), handle_msg);
-        let query_msg = QueryMsg::AllTokens {
-            viewer: Some(viewer.clone()),
-            start_after: None,
-            limit: Some(2),
-        };
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::TokenList { tokens } => {
-                let expected = vec!["NFT1".to_string(), "NFT2".to_string()];
-                assert_eq!(tokens, expected);
-            }
-            _ => panic!("unexpected"),
-        }
-
-        // test token supply public, with pagination starting after NFT2
-        let (init_result, mut deps) =
-            init_helper_with_config(true, true, true, false, true, false, true);
-        assert!(
-            init_result.is_ok(),
-            "Init failed: {}",
-            init_result.err().unwrap()
-        );
-
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT1".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My1".to_string()),
-                description: Some("Public 1".to_string()),
-                image: Some("URI 1".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg);
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT2".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My2".to_string()),
-                description: Some("Public 2".to_string()),
-                image: Some("URI 2".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT3".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My3".to_string()),
-                description: Some("Public 3".to_string()),
-                image: Some("URI 3".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT5".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My5".to_string()),
-                description: Some("Public 5".to_string()),
-                image: Some("URI 5".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let handle_msg = HandleMsg::MintNft {
-            token_id: Some("NFT4".to_string()),
-            owner: Some(HumanAddr("alice".to_string())),
-            public_metadata: Some(Metadata {
-                name: Some("My4".to_string()),
-                description: Some("Public 4".to_string()),
-                image: Some("URI 4".to_string()),
-            }),
-            private_metadata: None,
-            memo: None,
-            padding: None,
-        };
-        let _handle_result = handle(&mut deps, mock_env("admin", &[]), handle_msg); // test burn when status prevents it
-        let viewer = ViewerInfo {
-            address: alice.clone(),
-            viewing_key: "key".to_string(),
-        };
-        let query_msg = QueryMsg::AllTokens {
-            viewer: Some(viewer.clone()),
-            start_after: Some("NFT2".to_string()),
-            limit: Some(10),
-        };
-        let query_result = query(&deps, query_msg);
-        assert!(
-            query_result.is_ok(),
-            "query failed: {}",
-            query_result.err().unwrap()
-        );
-        let query_answer: QueryAnswer = from_binary(&query_result.unwrap()).unwrap();
-        match query_answer {
-            QueryAnswer::TokenList { tokens } => {
-                let expected = vec!["NFT3".to_string(), "NFT4".to_string(), "NFT5".to_string()];
-                assert_eq!(tokens, expected);
-            }
-            _ => panic!("unexpected"),
-        }
-    }
+    use crate::unittest_helpers::{extract_error_msg, init_helper_with_config};
 
     // test NftDossier query
     #[test]
